+++ rust/ommx/src/convert/state.rs
@@ -1,5 +1,4 @@
-use crate::{random::arbitrary_coefficient, v1::State};
-use proptest::prelude::*;
+use crate::v1::State;
 use std::collections::HashMap;
 
 impl From<HashMap<u64, f64>> for State {
@@ -22,22 +21,5 @@ impl IntoIterator for State {
 
     fn into_iter(self) -> Self::IntoIter {
         self.entries.into_iter()
-    }
-}
-
-impl Arbitrary for State {
-    type Parameters = (usize, u64);
-    type Strategy = BoxedStrategy<Self>;
-
-    fn arbitrary_with((size, max_id): Self::Parameters) -> Self::Strategy {
-        proptest::collection::hash_map(0..=max_id, arbitrary_coefficient(), 0..=size)
-            .prop_map(Self::from)
-            .boxed()
-    }
-
-    fn arbitrary() -> Self::Strategy {
-        (0..20_usize, 0..20_u64)
-            .prop_flat_map(Self::arbitrary_with)
-            .boxed()
     }
 }
